chore: add header alerts component
