+++ src/components/headerAlerts/HeaderAlerts.test.tsx
@@ -1,0 +1,59 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import { BrowserRouter as Router } from 'react-router-dom';
+import { findByShadowLabelText } from 'shadow-dom-testing-library';
+import { FlagsProvider, getFlags } from '../../flags';
+import { HeaderAlerts } from './HeaderAlerts';
+
+describe('HeaderAlerts', () => {
+  it('should render the header', async () => {
+    const { container } = render(
+      <FlagsProvider flags={getFlags()}>
+        <Router>
+          <HeaderAlerts pathname="/explore/api/appeals-status/docs" />
+        </Router>
+      </FlagsProvider>,
+    );
+
+    expect(
+      await findByShadowLabelText(
+        container,
+        /A new version of Appeals Status API (v1) will launch later this year./,
+      ),
+    ).toBeInTheDocument();
+  });
+
+  it('should render the header', async () => {
+    const { container } = render(
+      <FlagsProvider flags={getFlags()}>
+        <Router>
+          <HeaderAlerts pathname="/explore/api/va-facilities/docs" />
+        </Router>
+      </FlagsProvider>,
+    );
+
+    expect(
+      await findByShadowLabelText(
+        container,
+        /Version 1 of the VA Facilities API is launching soon./,
+      ),
+    ).toBeInTheDocument();
+  });
+
+  it('should render the header', async () => {
+    const { container } = render(
+      <FlagsProvider flags={getFlags()}>
+        <Router>
+          <HeaderAlerts pathname="/explore/api/veteran-confirmation/docs" />
+        </Router>
+      </FlagsProvider>,
+    );
+
+    expect(
+      await findByShadowLabelText(
+        container,
+        /Version 0 of the Veteran Confirmation API is deprecated and scheduled for deactivation/,
+      ),
+    ).toBeInTheDocument();
+  });
+});
